requirements.txt
playwright>=1.40.0
telebot>=0.0.5
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
numpy>=1.24.0
python-dotenv>=1.0.0
anthropic>=0.7.0